Adding InputStreamAndBase64string class.
Base64 is a text encoding scheme that provides portability for binary data between applications and platforms. Base64 can be used to store binary data in a database string column, thereby avoiding messy file operations. When combined with the data URI scheme, Base64 can be used to embed images in web pages and emails, in conformance with the HTML and Multipurpose Internet Mail Extensions (MIME) standards.
In this brief tutorial, we'll demonstrate Java Streaming IO functions and the built-in Java Base64 class to load binary data as an InputStream and then convert it to a String.

https://www.baeldung.com/java-inputstream-to-base64-string